+++ List/ICustomBuilder.cs
@@ -8,13 +8,13 @@ namespace FooProject.Collection
 {
     public interface ICustomBuilder<T>
     {
-        FixedList<T> CreateList(int init_capacity, int maxcapacity);
+        FixedList<T> CreateList(long init_capacity, long maxcapacity);
 
         LeafNode<T> CreateLeafNode();
 
         LeafNode<T> CreateLeafNode(T item);
 
-        LeafNode<T> CreateLeafNode(int count, FixedList<T> items);
+        LeafNode<T> CreateLeafNode(long count, FixedList<T> items);
 
         ConcatNode<T> CreateConcatNode(ConcatNode<T> node);
